Convert donation amount to float

+++ model.py
@@ -1,6 +1,6 @@
 import os
 
-from peewee import Model, CharField, IntegerField, ForeignKeyField
+from peewee import Model, CharField, FloatField, ForeignKeyField
 from playhouse.db_url import connect
 
 db = connect(os.environ.get('DATABASE_URL', 'sqlite:///my_database.db'))
@@ -13,7 +13,7 @@ class Donor(BaseModel):
     name = CharField(max_length=255, unique=True)
 
 class Donation(BaseModel):
-    value = IntegerField()
+    value = FloatField()
     donor = ForeignKeyField(Donor, field='name', backref='donations')
 
 class User(BaseModel):
